Create DelayedDelivery_README.md
This macro for Microsoft Outlook allows you to delay the delivery of an email to 6:00 AM, which is useful if you want to write an email outside of regular business hours but don't want to send it until the next morning. This macro can be easily customized to suit your specific needs.